requirements.txt
MySQL-python
pymongo

pycurl
lxml
grab
gevent
requests

pygeoip

south
django
django-debug-toolbar

djangorestframework
markdown
django-filter
